carpetas innecesarias eliminadas
